backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
python-dotenv==0.19.0
aiohttp==3.8.1
python-multipart==0.0.5
requests==2.26.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
torch==1.9.0
joblib==1.0.1 